Bigcommerce theme updates
theme 718 = updated ProductVideos.html and removed http:

theme 678 = This issue is part of Cart.js which was part of
https://blueprint-demo.mybigcommerce.com/ Is this issue present on this
theme? If not, could you please provide direction on what needs to
change?

theme 612 = A few panels/snippets have references like "<span>| <a
href="%%GLOBAL_ReviewLink%%#reviews">%%LNG_Next%%</a></span>" and the
section id is id="ProductReviews" - I updated the references to the
correct ID